+++ mapreader/__init__.py
@@ -6,12 +6,12 @@ from mapreader.download.sheet_downloader import SheetDownloader
 from mapreader.download.downloader import Downloader
 from mapreader.download.downloader_utils import create_polygon_from_latlons, create_line_from_latlons
 
-from mapreader.learn.load_annotations import AnnotationsLoader
-from mapreader.learn.datasets import PatchDataset
-from mapreader.learn.datasets import patchContextDataset
-from mapreader.learn.classifier import ClassifierContainer
-from mapreader.learn.classifier_context import classifierContext
-from mapreader.learn import custom_models
+from mapreader.classify.load_annotations import AnnotationsLoader
+from mapreader.classify.datasets import PatchDataset
+from mapreader.classify.datasets import patchContextDataset
+from mapreader.classify.classifier import ClassifierContainer
+from mapreader.classify.classifier_context import classifierContext
+from mapreader.classify import custom_models
 
 from mapreader.process import process
