+++ mapreader/classify/load_annotations.py
